deleted file check

+++ BitBoxSkeleton/src/main/java/unimelb/bitbox/ResponseHandler.java
@@ -38,7 +38,7 @@ public class ResponseHandler {
 
 			if (fManager.isSafePathName(pathName)) {
 				try {
-					if (!fManager.fileNameExists(pathName,fDesc.md5)) {
+					if (!fManager.fileNameExists(pathName)) {
 						//ensure no file in that path and try to create one
 						fManager.createFileLoader(pathName, desc.getString("md5"), desc.getLong("fileSize"), desc.getLong("lastModified"));
 						if (!fManager.checkShortcut(pathName)) {
@@ -50,7 +50,7 @@ public class ResponseHandler {
 							connection.sendCommand(JsonUtils.FILE_CREATE_RESPONSE(fDesc, pathName, "file exist already, create a copy from local", false));
 						}
 
-					} else {
+					} else if (!fManager.fileNameExists(pathName, fDesc.md5)){
 						//pathname already exists
 						// check for lastModified value
 						if(!fManager.deleteFile(pathName, fDesc.lastModified, fDesc.md5)) {
@@ -60,7 +60,7 @@ public class ResponseHandler {
 
 							fManager.createFileLoader(pathName, desc.getString("md5"), desc.getLong("fileSize"), desc.getLong("lastModified"));
 
-							if (fManager.checkShortcut(pathName)) {
+							if (!fManager.checkShortcut(pathName)) {
 								accept = true;
 								connection.sendCommand(JsonUtils.FILE_CREATE_RESPONSE(fDesc, pathName, "file loader ready", true));
 								connection.sendCommand(JsonUtils.FILE_BYTES_REQUEST(fDesc, d.getString("pathName"), 0, maximumBlockSize < fDesc.fileSize ? maximumBlockSize : fDesc.fileSize));
